refactor: move configuration code into a folder
