+++ mdbackup/config/__init__.py
@@ -1,0 +1,7 @@
+from mdbackup.config.configuration import Config, SecretConfig, StorageConfig
+
+__all__ = [
+    Config,
+    SecretConfig,
+    StorageConfig,
+]
